requirements.txt
streamlit>=1.31.0
crewai>=0.14.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
python-dotenv>=1.0.0
duckduckgo-search>=3.9.0 